+++ lib/src/showcase.dart
@@ -289,6 +289,12 @@ class Showcase extends StatefulWidget {
   /// Defaults to 0.5.
   final double scrollAlignment;
 
+  /// While target widget is out viewport then
+  /// whether enabling auto scroll so as to make the target widget visible.
+  /// This is used to override the [ShowCaseWidget.enableAutoScroll] behaviour
+  /// for this showcase.
+  final bool? enableAutoScroll;
+
   /// Highlights a specific widget on the screen with an informative tooltip.
   ///
   /// This widget helps you showcase specific parts of your UI by drawing an
@@ -350,6 +356,7 @@ class Showcase extends StatefulWidget {
   ///   - `tooltipActions`: A list of custom actions (widgets) to display within the tooltip.
   ///   - `tooltipActionConfig`: Configuration options for custom tooltip actions.
   ///   - `scrollAlignment`: Defines the alignment for the auto scroll function.
+  ///   - `enableAutoScroll`:This is used to override the [ShowCaseWidget.enableAutoScroll] behaviour for this showcase.
   ///
   /// **Assertions:**
   ///
@@ -407,6 +414,7 @@ class Showcase extends StatefulWidget {
     this.tooltipActions,
     this.tooltipActionConfig,
     this.scrollAlignment = 0.5,
+    this.enableAutoScroll,
   })  : height = null,
         width = null,
         container = null,
@@ -502,6 +510,7 @@ class Showcase extends StatefulWidget {
     this.tooltipActions,
     this.tooltipActionConfig,
     this.scrollAlignment = 0.5,
+    this.enableAutoScroll,
   })  : showArrow = false,
         onToolTipClick = null,
         scaleAnimationDuration = const Duration(milliseconds: 300),
@@ -580,7 +589,7 @@ class _ShowcaseState extends State<Showcase> {
     });
 
     if (activeStep == widget.key) {
-      if (showCaseWidgetState.enableAutoScroll) {
+      if (widget.enableAutoScroll ?? showCaseWidgetState.enableAutoScroll) {
         _scrollIntoView();
       }